+++ src/components/ui/Roster/RosterHeader.tsx
@@ -13,8 +13,8 @@ import { PopOverMenu } from './PopOverMenu';
 import { BaseProps, FocusableProps } from '../Base';
 
 export interface RosterHeaderProps extends BaseProps, FocusableProps {
-  /** The title of the roster header. */
-  title: string;
+  /** The title of the roster header, or an element that can render in this area */
+  title: string | ReactNode;
   /** The number of attendees in a meeing. */
   badge?: number;
   /** The string value to search in a roster. */